requirements.txt
# Pneumonia-XAI Requirements

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.7.0

# Metrics
scikit-learn>=1.2.0

# Vision Transformer
timm>=0.9.0

# Progress bar
tqdm>=4.65.0

# Optional - for development
# jupyter>=1.0.0
# ipywidgets>=8.0.0
